Valuation/apps/api/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
redis==5.0.8
weasyprint==61.2
pytest==8.3.3
pytest-django==4.9.0
requests==2.32.3
QuantLib==1.33
python-dotenv==1.0.1
